Add option to hide buckets from summaries (#314)
* Add option to hide buckets from summaries

This PR adds a new option to the bucket settings to exclude the bucket
from summaries on the report page. Previously large single transactions
which are not part of my monthly budget (like selling stocks) completly
messed up my statistics.

Im also changing how transfer transactions are displayed. They are now
hidden from the summaries to fix the statistics when the "input" and
"output" transactions of a transfer are in different months.
Discrepancies in the transfer bucket can still be seen in the "Data
Consistency" tab.

* Add Postgres Migration for BucketIsHiddenFromSummaries

* cleanup HiddenFromSummaries